requirements.txt
# Instagram API
instagrapi>=1.16.30

# Telegram Bot
python-telegram-bot==13.7

# Работа с изображениями
Pillow>=9.0.0

# База данных
SQLAlchemy>=1.4.0

# Многопоточность и асинхронность
# concurrent.futures - встроенный модуль, не требует установки
# asyncio - встроенный модуль, не требует установки

# Планировщик задач
schedule>=1.1.0
APScheduler>=3.6.3

# Прочие утилиты
python-dotenv>=0.19.0
requests>=2.26.0
loguru>=0.6.0

# Веб-драйверы (если используются для автоматизации)
selenium>=4.0.0
webdriver-manager>=3.5.0

# Обработка данных
pandas>=1.3.0

# Работа с файлами
moviepy>=1.0.3